requirements.txt
PyQt6==6.5.1
openai>=1.40.0
python-dotenv>=1.0.0